I could not figure out why the changes you made worked. Everything displays fine on my firefox and ie (if it goes funny can you send me your firefox version etc.. and i will have a play)
I tidied the code up and added a missing class statement that i must of deleted. (hope this is ok)

I rechecked that everything looked fine, the maximum number i can add without clipping is 250.00

I have left the attribute size declaration in so if needed the cell size can be forced large, or you can add a larger number in to the preset list.