references #107
I have made all of the code into functions, they need to have proper error functions added.

A lot of the functions should be moved to main as there are not global and pretain to the stats that are loaded on the myitcrm homepage.

Some code still needs sorting out, but that is next session.